Move poikkileikkaus information to jore database
